docs(svgs-maps): Layout of the H-8th floor (#59)
Values and parameters taken from the constructed svg map of the 8th floor of the hall building. Addition into the code of the rectangles representing classrooms and points of interest with coordinates from the svg file representing the floor map of the hall 8th floor. Rectangles are meant to be clickable objects.